+++ src/main/kotlin/com/coffee/product/repository/RoasteryRepository.kt
@@ -1,20 +1,21 @@
-package com.coffee.product.repository.roastery
+package com.coffee.product.repository
 
 import com.coffee.product.common.error.CustomException
 import com.coffee.product.common.error.type.ExceptionType
 import com.coffee.product.domain.roastery.Roastery
-import org.springframework.stereotype.Repository
+import org.springframework.data.jpa.repository.JpaRepository
 import java.util.UUID
 
-@Repository
-class RoasteryRepositoryImpl(
+interface RoasteryJpaRepository : JpaRepository<Roastery, UUID>
+
+class RoasteryRepository(
     private val roasteryJpaRepository: RoasteryJpaRepository,
-) : RoasteryRepository {
-    override fun register(roastery: Roastery): UUID {
+) {
+    fun register(roastery: Roastery): UUID {
         return roasteryJpaRepository.save(roastery).roasteryId
     }
 
-    override fun getRoasteryById(roasteryId: String): Roastery {
+    fun getRoasteryById(roasteryId: String): Roastery {
         return roasteryJpaRepository.findById(UUID.fromString(roasteryId))
             .orElseThrow {
                 throw CustomException(ExceptionType.ROASTERY_NOT_FOUND)
