--- src/main/kotlin/com/coffee/product/repository/roastery/RoasteryJpaRepository.kt
@@ -1,7 +1,0 @@
-package com.coffee.product.repository.roastery
-
-import com.coffee.product.domain.roastery.Roastery
-import org.springframework.data.jpa.repository.JpaRepository
-import java.util.UUID
-
-interface RoasteryJpaRepository : JpaRepository<Roastery, UUID>
