--- src/main/kotlin/com/coffee/product/repository/roastery/RoasteryRepository.kt
@@ -1,9 +1,0 @@
-package com.coffee.product.repository.roastery
-
-import com.coffee.product.domain.roastery.Roastery
-import java.util.UUID
-
-interface RoasteryRepository {
-    fun register(roastery: Roastery): UUID
-    fun getRoasteryById(roasteryId: String): Roastery
-}
