+++ src/main/kotlin/com/coffee/product/service/coffeebean/CoffeeBeanService.kt
@@ -4,5 +4,6 @@ import com.coffee.product.model.coffeebean.RegisterCoffeeBeanRequest
 
 interface CoffeeBeanService {
     fun registerCoffeeBean(registerCoffeeBeanRequest: RegisterCoffeeBeanRequest)
+
     fun getCoffeeBeanList(roasteryId: String)
 }
