+++ src/main/kotlin/com/coffee/product/service/coffeebean/CoffeeBeanServiceImpl.kt
@@ -2,14 +2,14 @@ package com.coffee.product.service.coffeebean
 
 import com.coffee.product.domain.roastery.Roastery
 import com.coffee.product.model.coffeebean.RegisterCoffeeBeanRequest
-import com.coffee.product.repository.coffeebean.CoffeeBeanRepository
-import com.coffee.product.repository.roastery.RoasteryRepository
+import com.coffee.product.repository.CoffeeBeanRepository
+import com.coffee.product.repository.RoasteryRepository
 import org.springframework.stereotype.Service
 
 @Service
 class CoffeeBeanServiceImpl(
     private val coffeeBeanRepository: CoffeeBeanRepository,
-    private val roasteryRepository: RoasteryRepository
+    private val roasteryRepository: RoasteryRepository,
 ) : CoffeeBeanService {
     override fun registerCoffeeBean(registerCoffeeBeanRequest: RegisterCoffeeBeanRequest) {
     }
